requirements.txt
Django==1.11.7
djangorestframework==3.7.3
psycopg2==2.7.3.2
gunicorn==19.7.1
whitenoise==3.3.1
dj-database-url==0.4.2